More rework tests.
